added an overflow check for max_body_size (#39)
* added an overflow check for max_body_size

* updated the variable type to long long everywhere

+++ src/Config/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: lyvan-de <lyvan-de@student.codam.nl>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/07 12:34:17 by lyvan-de          #+#    #+#             */
-/*   Updated: 2026/07/17 16:25:55 by lyvan-de         ###   ########.fr       */
+/*   Updated: 2026/07/23 17:06:14 by lyvan-de         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ class Server {
 	std::string	getHost() const;
 	std::string getRoot() const;
 	std::string getIndex() const;
-	size_t getMaxBodySize() const;
+	long long getMaxBodySize() const;
 	const std::map<int, std::string>& getErrorPages() const;
 	const std::vector<LocationBlock>& getLocationBlocks() const;
 	void setConfigDirectory(std::string configPath);
@@ -45,7 +45,7 @@ class Server {
 	std::string					_host;
 	std::string					_root;
 	std::string					_index;
-	size_t						_max_body_size;
+	long long					_max_body_size;
 	std::map<int, std::string>	_error_pages;
 	std::vector<LocationBlock>	_location_blocks;
 	bool						_finalized;
